xfs/293: tighten up checks for documented xfs_io commands
Some commands (like "zero") are simple words which commonly
occur in the manpage text even if they aren't documented as
commands.

Grep for "   $COMMAND" instead of the bare word, because
the documented commands show up as indented.

This reveals that the "zero" command is not documented yet.
(It catches "help" too, because it's documented differently;
I'll fix that up in the manpage when I add "zero").

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Eryu Guan <eguan@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>